// packages/shared/src/types.ts
export enum Zone {
  NEON_CITY = 'NEON_CITY',
  DESERT_STORM = 'DESERT_STORM',
  DEEP_OCEAN = 'DEEP_OCEAN',
  VOLCANO = 'VOLCANO',
  TUNDRA = 'TUNDRA',
  THE_VOID = 'THE_VOID',
  SPECIAL = 'SPECIAL',
}

export enum Rarity {
  COMMON = 'common',
  RARE = 'rare',
  EPIC = 'epic',
  LEGENDARY = 'legendary',
  MYTHIC = 'mythic',
}
// ...
export interface Warrior {
  id: string;
  name: string;
  displayName: string;
  zone: Zone;
  rarity: Rarity;
  basePrice: number;
  totalSupply: number;
  imageUrl: string;
  attack: number;
  defense: number;
  speed: number;
  magic: number;
  miningRate: number;
  lore?: string;
  ability?: string;
  abilityDesc?: string;
}